pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "envlint"
version = "0.1.0"
description = "Validate .env files against a schema. Catch missing and invalid environment variables before runtime."
readme = "README.md"
license = "GPL-3.0-or-later"
authors = [
    { name = "Kyle Cain", email = "kyle@kylecain.me" }
]
keywords = ["env", "environment", "validation", "dotenv", "schema", "cli", "devops", "ci-cd"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
    "Topic :: System :: Systems Administration",
]
requires-python = ">=3.9"
dependencies = [
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
envlint = "envlint.cli:app"

[project.urls]
Homepage = "https://github.com/cainky/envlint"
Repository = "https://github.com/cainky/envlint"
Issues = "https://github.com/cainky/envlint/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/envlint"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
